+++ examples/distributed/discovery_server.rs
@@ -19,10 +19,10 @@ pub struct DiscoveryServer {
 
 impl DiscoveryServer {
     pub fn new() -> Option<DiscoveryServer> {
-        let (event_queue, mut network) = Network::split();
+        let (mut network, event_queue) = Network::split();
 
         let listen_addr = "127.0.0.1:5000";
-        match network.listen(Transport::Tcp, listen_addr) {
+        match network.listen(Transport::FramedTcp, listen_addr) {
             Ok(_) => {
                 println!("Discovery server running at {}", listen_addr);
                 Some(DiscoveryServer { event_queue, network, participants: HashMap::new() })