+++ examples/file-transfer/sender.rs
@@ -11,11 +11,11 @@ enum Event {
 }
 
 pub fn run(file_path: &str) {
-    let (mut event_queue, mut network) =
+    let (mut network, mut event_queue) =
         Network::split_and_map(|net_event| Event::Network(net_event));
 
     let server_addr = "127.0.0.1:3005";
-    let (server_id, _) = match network.connect(Transport::Tcp, server_addr) {
+    let (server_id, _) = match network.connect(Transport::FramedTcp, server_addr) {
         Ok(server_id) => {
             println!("Connect to send by TCP at {}", server_addr);
             server_id