+++ src/adapters.rs
@@ -1,5 +1,6 @@
 pub mod template;
 pub mod tcp;
+pub mod framed_tcp;
 pub mod udp;
 pub mod web_socket;
 // Add new adapters here